+++ projects/ausmittlung-lib/src/lib/components/contest-detail/contest-majority-election-detail/contest-majority-election-result-input/contest-majority-election-result-input.component.ts
@@ -27,6 +27,9 @@ export class ContestMajorityElectionResultInputComponent {
   public showInvalidVotes: boolean = false;
 
   @Input()
+  public showEmptyVotes: boolean = true;
+
+  @Input()
   public readonly: boolean = true;
 
   @Output()