+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# ✨ Changelog (`v1.33.1`)
+# ✨ Changelog (`v1.35.1`)
 
 All notable changes to this project will be documented in this file.
 
@@ -8,11 +8,71 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Version Info
 
 ```text
-This version -------- v1.33.1
-Previous version ---- v1.25.0
+This version -------- v1.35.1
+Previous version ---- v1.33.1
 Initial version ----- v1.25.0
-Total commits ------- 31
+Total commits ------- 11
 ```
+
+## [v1.35.1] - 2022-11-16
+
+### 🔄 Changed
+
+- fix mail voting channel label
+
+## [v1.35.0] - 2022-11-16
+
+### 🔒 Security
+
+- configure client refresh token flow (rfc-6749)
+
+## [v1.34.4] - 2022-11-03
+
+### 🆕 Added
+
+- add eVoting write in mapping to invalid ballot
+
+## [v1.34.3] - 2022-10-31
+
+### 🆕 Added
+
+- add result state change listener for erfassung
+
+## [v1.34.2] - 2022-10-31
+
+### 🔄 Changed
+
+- update can set state on result after on init
+
+## [v1.34.1] - 2022-10-28
+
+### 🔄 Changed
+
+- set all results to audited tentatively depending responsible tenant
+
+## [v1.34.0] - 2022-10-27
+
+### 🆕 Added
+
+- Reset counting circle results in testing phase
+
+## [v1.33.4] - 2022-10-14
+
+### 🔄 Changed
+
+- Fixed majority election lot decision typo
+
+## [v1.33.3] - 2022-10-13
+
+### 🔄 Changed
+
+- no empty vote count and no invalid vote count for single mandate
+
+## [v1.33.2] - 2022-10-04
+
+### 🔄 Changed
+
+- Updated voting-library to fix layouting issues
 
 ## [v1.33.1] - 2022-09-28